Migrated waiting room into exams router
