+++ api/src/main.py
@@ -51,14 +51,13 @@ app.add_middleware(
 )
 
 # Include routers
-from src.routers import user, subject, topic, question, question_option, exam, waiting_room
+from src.routers import user, subject, topic, question, question_option, exam
 app.include_router(user.router, prefix="/api/users", tags=["users"])
 app.include_router(subject.router, prefix="/api/subjects", tags=["subjects"])
 app.include_router(topic.router, prefix="/api/topics", tags=["topics"])
 app.include_router(question.router, prefix="/api/questions", tags=["questions"])
 app.include_router(question_option.router, prefix="/api/question-options", tags=["question-options"])
 app.include_router(exam.router, prefix="/api/exams", tags=["exams"])
-app.include_router(waiting_room.router, prefix="/api/waiting-rooms", tags=["waiting-rooms"])
 
 @app.get("/health")
 async def health_check():
